fix(alerts): address review — auth bypass, middleware order, and layout
Two real defects found by review, plus the file layout asked for.

**Empty decrypted key authenticated every request.** `validate()` ran against
`Settings<SecuredSecret>` — under a KMS backend that is a *handle*, not the key.
A well-formed handle resolving to an empty string passed the boot check, and an
empty `X-Internal-Api-Key` header then compared equal to the empty configured
key. Re-validate after decryption, so the check happens on both sides of the
`SecuredSecret` -> `RawSecret` transition.

**Root spans recorded an empty request id.** actix runs the last registered
`.wrap` first, so registering `RequestIdentifier` before `TracingLogger` made
the logger outermost; `CustomRootSpanBuilder` read the id from extensions
before the identifier had inserted it and fell back to `unwrap_or_default()`.
Swapped to match the router at `router/src/lib.rs:574`. Verified: the span's
request id now matches the one in the rejection log line.

**Layout now follows the router.** `core/` decides, `routes/` exposes:

  routes.rs, routes/app.rs        — the whole route tree, one file
  routes/health_check.rs          — health handler
  routes/alerts.rs                — ping handler
  core.rs, core/notifier.rs       — the domain seam (was src/notifier.rs)

`routes/app.rs` means "what does this expose, and what guards it?" is one file
rather than a search.

Also: `InternalServerError` ordered first in `AlertsError`, `ApiErrorResponse`
and their matches; imports hoisted to module top in `errors/actix.rs`.

Re-verified: `/health` 200 unauthenticated; `/alerts/ping` 401 with no key,
with a wrong key, and with an empty key header; 200 with a valid key. Clippy
clean.

Ref: juspay/hyperswitch-cloud#23114
